feat: InQuirer API
