+++ routes/web.php
@@ -19,14 +19,3 @@ use Illuminate\Support\Facades\Auth;
 |
 */
 
-Auth::routes();
-Route::get('/', [BookController::class, 'index'])->name('book.index');
-Route::post('/search', [BookController::class, 'search'])->name('book.search');
-Route::get('/book/{id}', [BookController::class, 'show'])->name('book.show');
-Route::get('rental/book', [RentalController::class, 'mybook'])->name('rental.book');
-Route::get('rental/detail', [RentalController::class, 'mydetail'])->name('rental.detail');
-Route::resource('category', CategoryController::class);
-Route::resource('bookcart', BookcartController::class);
-Route::resource('rental', RentalController::class);
-Route::resource('area', AreaController::class);
-Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
